+++ app/Database/Repositories/Mongo/PagamentoRepository.js
@@ -1,8 +1,12 @@
 // PagamentoRepository.js
 const mongoose = require('mongoose');
-const Pagamento = mongoose.model('Pagamento');
+const Pagamento = require('../../Models/Pagamento');
+const AbstractRepository = require('../AbstractRepository');
 
-class PagamentoRepository {
+class PagamentoRepository extends AbstractRepository 
+{
+  _objectIdRegex = /^[a-f\d]{24}$/i; // regex para id
+
   async criar(pagamento) {
     try {
       const novoPagamento = new Pagamento(pagamento);
@@ -15,8 +19,19 @@ class PagamentoRepository {
 
   async listar() {
     try {
-      const pagamentos = await Pagamento.find().exec();
+      const pagamentos = await Pagamento.find();
       return pagamentos;
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  async encontrar(id) {
+    try {
+      if(! this._objectIdRegex.test(id)) throw new Error('Informe um id valido');
+      const pagamento = await Pagamento.findById(id);
+      if (!pagamento) throw new NotFoundException('pagamento não encontrado');
+      return pagamento;
     } catch (error) {
       throw error;
     }
@@ -24,6 +39,7 @@ class PagamentoRepository {
 
   async atualizar(id, dados) {
     try {
+      if(! this._objectIdRegex.test(id)) throw new Error('Informe um id valido');
       const pagamentoAtualizado = await Pagamento.findByIdAndUpdate(id, dados, { new: true });
       return pagamentoAtualizado;
     } catch (error) {
@@ -33,6 +49,7 @@ class PagamentoRepository {
 
   async deletar(id) {
     try {
+      if(! this._objectIdRegex.test(id)) throw new Error('Informe um id valido');
       await Pagamento.findByIdAndDelete(id);
     } catch (error) {
       throw error;
@@ -40,4 +57,4 @@ class PagamentoRepository {
   }
 }
 
-module.exports = PagamentoRepository;
+module.exports = new PagamentoRepository();
